rename eventManager to custom hook manner
