+++ frontend/src/hooks/useEventManagerContract.ts
